pyproject.toml
[project]
name = "voice-agent-tests"
version = "0.1.0"
description = "MVP for testing ASR providers"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.1",
    "python-dotenv>=1.0.0",
    "openai>=1.50.0",
    "aiohttp>=3.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

